Move conflicts, build policy and index routing out of config.py
