+++ nab-project/benchmarks/universal_scenarios.py
@@ -49,7 +49,8 @@ from universal_result import (
 )
 
 from nab_index.urllib3_async_transport import Urllib3AsyncTransport
-from nab_project.config import NabProjectConfig, enforce_build_policy_for_targets
+from nab_project.build_policy import enforce_build_policy_for_targets
+from nab_project.config import NabProjectConfig
 from nab_project.fetch import FetchCoordinator
 from nab_project.lockfile import build_pylock, render_lock
 from nab_project.resolve import (
